server/requirements.txt
zhipuai
flask_cors
flask
pymysql
python-dotenv
